+++ jim-file.c
@@ -49,9 +49,11 @@
 #include <stdio.h>
 #include <errno.h>
 #include <sys/stat.h>
+#include <time.h>
 
 #include <jimautoconf.h>
 #include <jim-subcmd.h>
+#include "jim-win32compat.h"
 
 #ifdef HAVE_UTIMES
 #include <sys/time.h>
@@ -218,10 +220,10 @@ static int file_cmd_dirname(Jim_Interp *interp, int argc, Jim_Obj *const *argv)
     }
     else if (ISWINDOWS && p[-1] == ':') {
         /* z:/dir => z:/ */
-        Jim_SetResultString(interp, path, p - path + 1);
+        Jim_SetResultString(interp, path, (int)(p - path + 1));
     }
     else {
-        Jim_SetResultString(interp, path, p - path);
+        Jim_SetResultString(interp, path, (int)(p - path));
     }
     return JIM_OK;
 }
@@ -236,7 +238,7 @@ static int file_cmd_rootname(Jim_Interp *interp, int argc, Jim_Obj *const *argv)
         Jim_SetResult(interp, argv[0]);
     }
     else {
-        Jim_SetResultString(interp, path, p - path);
+        Jim_SetResultString(interp, path, (int)(p - path));
     }
     return JIM_OK;
 }
@@ -349,7 +351,7 @@ static int file_cmd_join(Jim_Interp *interp, int argc, Jim_Obj *const *argv)
 
     /* Probably need to handle some special cases ... */
 
-    Jim_SetResult(interp, Jim_NewStringObjNoAlloc(interp, newname, last - newname));
+    Jim_SetResult(interp, Jim_NewStringObjNoAlloc(interp, newname, (int)(last - newname)));
 
     return JIM_OK;
 }
